refactor(bindings): extract PA + HLDA into src/python/hierarchical.rs (#385)
First incremental split of the ~18k-line src/python/mod.rs. Move the two
hierarchy models — Pachinko Allocation (PA) and hierarchical LDA (HLDA), a
cohesive pair at the end of the file — into a new src/python/hierarchical.rs,
following the established per-module recipe (a directory module pulled in with
`use super::*`, e.g. nmf_lsa.rs). Their PA/HLDA-exclusive serde save-structs
(PaState/HldaState) move with them; shared helpers, type aliases, and the
save-format tags stay in mod.rs and are reached via `use super::*`.

No public API change: both classes are still registered in the same #[pymodule]
fn (`use hierarchical::{HLDA, PA}`) and imported as topica.PA / topica.HLDA. Also
update the #381 IMPL map so their binding column points at hierarchical.rs, and
regenerate the model map.

mod.rs: 18407 -> 17658 lines. Gates green: cargo check (default + python), clippy
-D warnings, fmt, preflight (gen-check + stub sync), and the full pytest suite
(3321 passed, incl. PA/HLDA gold, save/load, conformance, registry).
