+++ python/topica/registry.py
@@ -279,8 +279,8 @@ IMPL: dict[str, ImplInfo] = {
     "BTM": _i("src/btm.rs", "src/python/btm.rs", "collapsed Gibbs over biterms", "", "parity/btm_compare.py, tests/test_btm.py"),
     "DTM": _i("src/dtm.rs", "src/python/mod.rs", "variational Kalman over time slices", "", "parity/dtm_gold.py"),
     "DETM": _i("src/detm.rs", "src/python/mod.rs", "embedding VAE + LSTM q(eta) (etm_vae.rs)", "", "parity/detm_gold.py"),
-    "HLDA": _i("src/hlda.rs", "src/python/mod.rs", "nested-CRP collapsed Gibbs", "", "parity/hlda_gold.py"),
-    "PA": _i("src/pa.rs", "src/python/mod.rs", "collapsed Gibbs over a topic DAG", "", "parity/pa_gold.py"),
+    "HLDA": _i("src/hlda.rs", "src/python/hierarchical.rs", "nested-CRP collapsed Gibbs", "", "parity/hlda_gold.py"),
+    "PA": _i("src/pa.rs", "src/python/hierarchical.rs", "collapsed Gibbs over a topic DAG", "", "parity/pa_gold.py"),
     "BERTopic": _i("src/bertopic.rs", "src/python/mod.rs", "embedding clustering (cluster.rs, reduce.rs, represent.rs)", "embeddings", "parity/bertopic_gold.py"),
     "Top2Vec": _i("src/top2vec.rs", "src/python/mod.rs", "embedding clustering (cluster.rs, reduce.rs)", "embeddings", "parity/top2vec_gold.py, parity/top2vec_compare.py"),
     "ETM": _i("src/etm.rs", "src/python/mod.rs", "variational EM over word embeddings (ctm.rs)", "", "parity/etm_gold.py"),
